+++ Brio/Game/Types/FacewearTypes.cs
@@ -17,17 +17,17 @@ public partial class FacewearUnion : OneOfBase<Glasses, None>
     }
 }
 
-public record struct FacewearId(byte Id)
+public record struct FacewearId(ushort Id)
 {
     public static FacewearId None { get; } = new(0);
 
     public static implicit operator FacewearId(DyeUnion dyeUnion) => dyeUnion.Match(
-        dyeRow => new FacewearId((byte)dyeRow.RowId),
+        dyeRow => new FacewearId((ushort)dyeRow.RowId),
         none => None
     );
 
-    public static implicit operator FacewearId(int dye) => new((byte)dye);
-    public static implicit operator FacewearId(byte dye) => new(dye);
-    public static implicit operator byte(FacewearId id) => id.Id;
+    public static implicit operator FacewearId(int dye) => new((ushort)dye);
+    public static implicit operator FacewearId(ushort dye) => new(dye);
+    public static implicit operator ushort(FacewearId id) => id.Id;
     public static implicit operator int(FacewearId id) => id.Id;
 }
